iron-dome/requirements.txt
psutil
watchdog
python-magic-bin;python_version < "3.8"
python-magic;python_version >= "3.8"